requirements.txt
# Python dependencies for Render deployment
# Generated from pyproject.toml

fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-dotenv>=1.0.0
httpx>=0.27.0
pydantic>=2.9.0
supabase>=2.0.0
gunicorn>=21.0.0
stripe>=14.0.0
python-multipart>=0.0.6
